pyproject.toml
[tool.poetry]
name = "jasmin-services"
version = "0.1.0"
description = "Django application for managing services, roles and access."
authors = ["Matt Pryor <matt.pryor@stfc.ac.uk>"]
license = "BSD-3-Clause"

[tool.poetry.dependencies]
python = "^3.6.8"

# Dependencies of dependencies. Need to specify here so poetry knows where to get them from.
jasmin-ldap = {git = "https://github.com/cedadev/jasmin-ldap.git", tag = "v1.0.0"}

Django = "^3.2"
django-bootstrap5 = "^21.2"
django-markdown-deux = "^1.0.5"
python-dateutil = "^2.8.2"
django-polymorphic = "^3.1.0"
requests = "^2.27.1"
jasmin-ldap-django = {git = "https://github.com/cedadev/jasmin-ldap-django.git", tag = "v1.0.1"}
jasmin-metadata = {git = "https://github.com/cedadev/jasmin-metadata.git", tag = "v1.0.0"}
jasmin-notifications = {git = "https://github.com/cedadev/jasmin-notifications.git", tag = "v1.0.0"}
django-countries = "^7.2.1, !=7.3.0"
djangorestframework = {version = "^3.13.1", optional = true}
django-filter = {version = "^21.1", optional = true}
jasmin-django-utils = {git = "https://github.com/cedadev/jasmin-django-utils.git", tag = "v1.0.2", optional = true}

[tool.poetry.extras]
api = ["djangorestframework", "django-filter", "jasmin-django-utils"]

[tool.poetry.dev-dependencies]
tox = "^3.24.5"
pytest = "^7.0.1"
black = "^22.1.0"
isort = "^5.10.1"

[tool.black]
line-length = 88
target-version = ['py36','py38']

[tool.isort]
profile = "black"

[tool.mypy]
strict = false

[build-system]
requires = ["poetry-core>=1.0.0", "setuptools>=59.6.0"]
build-backend = "poetry.core.masonry.api"
